requirements.txt
aiodns==2.0.0
aiohttp==3.7.4
aiohttp-socks==0.5.5
appnope==0.1.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.4
dateparser==1.0.0
decorator==4.4.2
elasticsearch==7.11.0
fake-useragent==0.1.11
geographiclib==1.50
geopy==2.1.0
googletransx==2.4.2
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
multidict==5.1.0
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycares==3.1.1
pycparser==2.20
Pygments==2.8.0
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-socks==1.2.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
schedule==1.0.0
six==1.15.0
soupsieve==2.2
torpy==1.1.5
traitlets==5.0.5
twint @ git+https://github.com/himanshudabas/twint.git@dc45e7de5eb2a2f63b5c2b84d30da846d0bcbf5b
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
yarl==1.6.3
